Refactored code base into multiple files to make it easier to spread work across project.
