+++ js/home.js
@@ -1,0 +1,8 @@
+'use strict';
+
+
+function renderHome(request, response) {
+  response.render('index');
+}
+
+module.exports = renderHome;
