+++ js/search.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function searchCocktails(request, response) {
+    response.render('search/search');
+}
+
+module.exports = searchCocktails;
